[UPDATED] Refactored project structure and renamed components for consistency
