+++ backend/src/main/java/com/bvm/tic_tac_toe/TicTacToeApplication.java
@@ -1,13 +1,13 @@
-package com.bvm.tik_tak_toe;
+package com.bvm.tic_tac_toe;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class TikTakToeApplication {
+public class TicTacToeApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(TikTakToeApplication.class, args);
+		SpringApplication.run(TicTacToeApplication.class, args);
 	}
 
 }
