+++ backend/src/main/java/com/bvm/tic_tac_toe/exceptions/body/Commence.java
@@ -1,4 +1,4 @@
-package com.bvm.tik_tak_toe.exceptions.body;
+package com.bvm.tic_tac_toe.exceptions.body;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
