+++ backend/src/main/java/com/bvm/tic_tac_toe/exceptions/exception/InvalidMoveException.java
@@ -1,0 +1,7 @@
+package com.bvm.tic_tac_toe.exceptions.exception;
+
+public class InvalidMoveException extends RuntimeException {
+    public InvalidMoveException(String message) {
+        super(message);
+    }
+}
