+++ backend/src/main/java/com/bvm/tic_tac_toe/exceptions/exception/NoSuchGameFoundException.java
@@ -1,4 +1,4 @@
-package com.bvm.tik_tak_toe.exceptions.exception;
+package com.bvm.tic_tac_toe.exceptions.exception;
 
 public class NoSuchGameFoundException extends RuntimeException {
     public NoSuchGameFoundException(String message) {
