+++ backend/src/main/java/com/bvm/tic_tac_toe/services/GameService.java
@@ -1,0 +1,18 @@
+package com.bvm.tic_tac_toe.services;
+
+import com.bvm.tic_tac_toe.exceptions.exception.InvalidMoveException;
+import com.bvm.tic_tac_toe.exceptions.exception.NoSuchGameFoundException;
+import com.bvm.tic_tac_toe.model.GameMove;
+import com.bvm.tic_tac_toe.model.GameState;
+
+public interface GameService {
+    GameState createGame();
+
+    GameState joinGame(String gameId);
+
+    void endGame(String gameId) throws NoSuchGameFoundException;
+
+    GameState getGame(String gameId) throws NoSuchGameFoundException;
+
+    GameState updateGame(String gameId, GameMove move) throws InvalidMoveException, NoSuchGameFoundException;
+}
