--- backend/src/main/java/com/bvm/tik_tak_toe/exceptions/exception/InvalidMoveException.java
@@ -1,7 +1,0 @@
-package com.bvm.tik_tak_toe.exceptions.exception;
-
-public class InvalidMoveException extends RuntimeException{
-    public InvalidMoveException(String message) {
-        super(message);
-    }
-}
