+++ backend/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=tik-tak-toe
+spring.application.name=tic-tac-toe
 
 # the depug level of the application
 #logging.level.org.springframework=DEBUG
