+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "tiktaktoe",
+  "name": "tictactoe",
   "private": true,
   "version": "0.0.0",
   "type": "module",
